requirements.txt
mysql-connector-python==8.0.17
SQLAlchemy==1.3.18
requests==2.21.0
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-Cors==3.0.7
pytest==4.3.1
PyMySQL==0.9.3
pytest-cov==2.6.1
pytest-mock==1.10.2
gunicorn
